Add Mypage function 作成途中ですが、一度pushします。 Pushed by AKI
