--- lib/screens/mypage/mypage_screen.dart
@@ -1,16 +1,0 @@
-import "package:flutter/material.dart";
-
-class MyPageScreen extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        centerTitle: true,
-        backgroundColor: Colors.lightBlue,
-      ),
-      body: Center(
-        child: Text("MyPage Screen"),
-      ),
-    );
-  }
-}
