+++ lib/utils/constans.dart
@@ -1,0 +1,4 @@
+enum ProfileMode{
+  MYSELF,
+  OTHER,
+}
